Bump to flask v1.0.2
 ## Summary
This brings our Flask version up to 1.0.2. By not sending `SERVER_NAME`
as the Cookie Domain back in responses in development environments, we
can avoid the warning that Flask generates around cookies not being
valid for domains without at least two components (i.e. one `.`
separator).

The other option is that we all run Flask locally via a dotted domain,
e.g. `local.host` and add this to our hosts file. Unfortunately it
doesn't seem possible to get Heroku to add an entry to its own host file
when under test, so we would still be running all our tests with the
`localhost` domain. It feels better to keep dev+test more closely
aligned, so have opted for this implementation. Open to suggestions.

 ## Ticket
https://trello.com/c/aH8nJmM5/1107