Raise build exceptions to stderr
 ## Summary
If a build fails, our process captures the exception and supresses
it, injecting the stacktrace into the relevant build record in the
database. This record is then scraped later on by an hourly
cronjob, which then emails devs.

When a build fails locally, there's no significant indication that
anything went wrong - you have to always remember to look in the
database to see if it succeeded. This has caught us out a few times.

We want build errors to be reported immediately to stdout/stderr,
which can then be monitored by sentry to provide immediate alerts
on build failures. It also helps testing local builds as the error
isn't hidden.

 ## Ticket
https://trello.com/c/5KZOCGf7/1109