Do not submit form for arbitrary link clicks; use green button to submit
For some unspecified reason (I guess the rationale was so people don't
lose things they have typed when they click a link?) the regular GET
links to add or edit charts and tables also first submitted the form
on the page.

This is really weird and broken - let's just use the green button for
submitting the form, like it's designed for.