Add a flag to update_dimension for whether to update_classification
We currently have three different POST routes that eventually call
`update_dimenson()`:
 1. the "edit dimension" metadata page (title, etc.)
 2. save chart endpoint
 3. save table endpoint

The problem is that if we run the reclassification when saving the
metadata then existing dimension-level classifications can end up
being deleted if no associated Chart or Table classification has yet
been created.

We only want to call the reclassification if a chart or table has been
updated, so need some way to signal this to update_dimension().