Add custom classification selector to charts
 ## Summary
The HTML for selecting/defining the custom classification of data, if
the user is adding data that either does not fit one of our recognised
classifications, or is wrongly classified, is only present on table
builder. This patch pulls it out into a separate shared fragment that is
included in the relevant places in both chart- and tablebuilder.

 ## Ticket
https://trello.com/c/r1V4liys/1028