Save id from internal classification link, not from web form
This fixes a bug where the "external" classifcation codes were being passed to be saved on chart or table, rather than the id from the "internal" classificatin_link.  In most cases this is fine, but in cases where there is a "+" on the end of the external classification the request to save would fail because we don't store the trailing "+" internally.

The culprit lines of code were:
```
chart.classification_id = code_from_builder
...
table.classification_id = code_from_builder
```

Also refactored a bit to reduce some duplication of code.  This feels like it could still be done better by pushing at least some of the functionality here back to the ChartAndTableMixin on the model, but I couldn't find a neat way to do it and this is better than what we had previously.