Clean DB before tests, not after
 ## Summary
At the moment, we tear down the data from the test database at the end
of a function that uses a db_session. This is problematic as some of our
migrations inject data into the database, which might then conflict with
what tests expected. This happens to work fine if you are running the
full test suite, as by the time you get to the functions that have
conflicting data, the pre-existing data has been removed. But if you try
to run a single specific tests (e.g.
test_can_reject_a_measure_in_review_as_editor), then you get a PK
conflict error as the test stubs try to insert a record that conflicts
with data inserted by the migrations.

By removing the data before tests that use the DB, we can prevent this
problem. A side-effect is that the test DB might be left in a 'dirty'
state at the end of test runs. This feels acceptable, and possibly even
preferrable, so that you can easily look into the database after a
failed test to see the state it's at.