Remove manual scrolling
 ## Summary
Our Heroku tests have been broken for about a week because of an
inability to get past a specific Selenium test - specifically one that
creates a new measure and then tries to set a publisher on it. Upon
sending it to review, a validation error occurred because the select
option for 'Department of Work and Pensions' had not been correctly
selected by the driver.

Removing these random 8 presses of the UP key seem to allow Heroku's
chromedriver to confirm the option selection.

This isn't a particularly _satisfying_ fix, because it's very difficult
to validate **why** this was happening, and indeed that it **has** been
definitively fixed. But it looks like it's fixed us (for now)...

 ## Ticket
https://trello.com/c/qDRWxakt/1120