Clear text input manually
 ## Summary
We think that selenium's chromedriver `element.clear()` does not work
correctly on some Heroku instances. It looks like doing a CTRL+A, DEL
within the element might clear it successfully. This might be because
the manual key-based method triggers the appropriate JS events for our
autocomplete text input, whereas `.clear()` does not.

 ## Ticket
https://trello.com/c/qDRWxakt/1120