Fix static site build
 ## Summary
https://github.com/racedisparityaudit/rd_cms/pull/803 renamed a number
of variables used in templates, but didn't update render_template
arguments in tests, so this slipped in. I've run the static site build
locally and it now passes.

I'm going to look into how we might test the static site build process
so that these sort of bugs don't creep in, but getting this in now
should unblock the pipeline.