Rudimentary test for static site build
 ## Summary
I've merged a few commits over the last months that have broken static
site builds. This is, in part, due to a lack of testing around this
process. This patch introduces a rudimentary test for the static site
build process - it will generate a set of pages (homepage, topic page,
measure page, dashboards, 'Ethnicity in the UK' static pages) and send
them to /dev/null. The main thing this tests is that the core templates
for the static site can be called correctly and return without raising
errors. This is not a thorough test for the site builder, but it should
give us _some_ level of assurance, where previously there was none
(beyond manual QA).

 ## Ticket
https://trello.com/c/7Z2PQnb1/1138