Fix static site links with trailing slashes
 ## Summary
Some of our routes have trailing slashes and others don't (for example,
`/users/` vs `/admin`. This gives us inconsistent URLs and makes it
harder to predict what structure we need when we export the site to
static files and upload it to S3. This PR moves us to having **no
trailing slashes** by default - though also disables Flask's strict
handling of slashes - so if a trailing slash is provided, it will
redirect to the non-trailing version with a 302. The trailing slash
historically implied a directory-like structure, which isn't really what
we have. We have a set of browsable pages, so sans-slash feels fine.

Given that we now also serve our static files from an S3 bucket which is
enabled for web hosting, it will automatically resolve something like
`/users` down to `/users/index.html`, which lets us remove the extra
step in the s3_deployer that moves index files up a directory level and
flattens them out into just `/users`. This simplifies our build process
and gives us a slightly more standard file layout.

 ## Research
https://webmasters.googleblog.com/2010/04/to-slash-or-not-to-slash.html
https://www.seoblog.com/2018/05/remove-trailing-slash-urls/
https://stackoverflow.com/questions/33241050/trailing-slash-triggers-404-in-flask-path-rule

 ## Deployment process
When this goes out, some objects in the S3 bucket will no longer get
updated, but could potentially get served if the correct URL is hit. We
will need to manually remove these keys so that S3/CloudFront can start
serving the new ones and redirecting as appropriate. These keys will
mainly be at the top level(s) of the bucket, where keys have been
translated from a/b/c/index.html to a/b/c (or whatever).

 ## Ticket
https://trello.com/c/0J4BvUYA/1108