Update URL construction parameters
 ## Summary
We renamed the 'topic' and 'subtopic' variables to 'topic_uri' and
'subtopic_uri', which made these poorly-generated URLs fail to point to
valid paths - or error out. This fixes the naming error and adds some
tests that will help ensure these links continue to work in the future.
Unfortunately, due to the way the measure pages are rendered by
for the static build, we aren't able to easily check that the correct
links are generated for the live site. This is because the static site
builder renders measure pages manually and passes in specially-formatted
dimension objects, which have extra keys not normally present on a
dimension object. The template expects these to be available for all
measure renders in static mode - but using the `static_mode` query param
to requests, which _should_ force a page to render in static mode, does
not patch the dimension objects in the same way. A tech improvement
ticket has been created to address this in the future.

 ## Ticket
https://trello.com/c/HAtcbSNQ/1139