Show same download description on CMS and live
 ## Summary
This commit changes how we present the filename/title for the measure
page source data download. At the moment, it is different depending on
whether the user is on the publisher or the live site. On the
publisher, we display a 'human-readable' title for the file download
and file type. On the live site, we display the raw filename for the
download. This feels backwards. After speaking to Jags and Amy, we've
agreed to consolidate this and only display the 'human-readable' file
title and type on both platforms.