Show classification short name in drop-down but long on dimension page
The classification "short names" should be shown in the drop-down to
pick Custom classifications as the worry is that if the only difference
is a value at the end of a long list then users won't be able to tell
which option is which.

But on the Dimension page itself we show the full list of ethnicities.