Order pages in the database
 ## Summary
We have some helper methods on `page_service` that return a list of
pages from the database. Some of these methods return pages in an
undefined order because they do not specify an order clause when
querying the database. This makes it easy for developers to forget to
order the results manually.

The safest solution here is to order the `.all()` Page queries
explicitly - and once only - in the page service. All callers can then
rely on it being in a specific order. If they need it in a different
order, they can sort it manually.

This lack of ordering caused a bug where some measure pages were
incorrectly showing the "There is a newer version of this measure"
banner at the top of the page, because we were simply grabbing the first
measure page from the result set and expected it to be the latest
version.

 ## Ticket
https://trello.com/c/hg97gPCJ/1151