Rename things in DimensionService
This drops all mention of "link", "external" and "internal".

"External" things have been renamed as "builder...", as all these things
come from chart and table builders.

"Internal" things are just Classifications (i.e. from the Classification
model), and so are now named accordingly.

"Link builder" is renamed as "Matcher" because it matches things from
the chart and table builder to things in the database.