Update outdated test requirements
These aren't insecure but are outdated.

Updates all test requirements to latest version, *except* Selenium,
which we have pinned to a version known to work on Heroku.

Also removed "Click" from test requirements as it is in the main
requirements file too anyway (and is a dependency of one of the packages
there).