Do not fail entire test run until all tests have been run
Sometimes you don't care that the code style failed but just want
to see your python tests run in CI environment.

This change means the test run will still be flagged as a fail,
but only after everything has run.

The downside to this is that we'll have to wait longer for feedback
from bad CI builds in the build channel.  I'm not sure if this
outweighs the upside of being able to see full test suite output
for the broken build - for discussion.