Rollback pytest to 3.4.0
Updating to the latest pytest resulted in the functional tests failing
in our CI environment on Heroku - but not locally - with the error:
`psycopg2.OperationalError: SSL error: decryption failed or bad record mac`

Rolling back the version of pytest and related dependencies until we can
figure out what the issue is.

NOTE: have tried pytest 3.6.0 and that is broken with the same error.