Make statc site build test do more
Previously when test_static-site_build() ran there wasn't the right
fixture data to prompt the build to descend into all code paths (e.g.
several of the dashboard builds were omitted because there was no data
in the relevant materialized view that drives the dashboards).

I have verified through liberal and rigorous use of `print()` statements
throughout the build.py and data_helpers.py files that the combination
of fixtures, publishing a page and refreshing the views contained in
this commit forces the test of the build to:
* Build the homepage, a topic and a subtopic
* Build a latest version measure page
* Build an old version of a measure page
* Create all dashboard directories
* Build the Ethnicities dashboard with a page linked from it
* Build the Ethnicity Classifications dashboard with a page linked from it
* Build the Locations dashboard with a page linked from it
* Build a bunch of static pages