Remove "analytics" folder
It's not completely clear what this was ever useful for.

With a bit of jiggery-pokery the SQL queries can be coerced into
running against our current databse structure to produce... something.

It *might* be a table with one row for every cell in every table on
every page on the website.

Which is quite a feat. If it's right.

But...

I can't see why we would ever need to do this, and if we did then we
probably want to re-write the queries anyway so we understand what
they're doing.

And as for the report... in RMarkdown... I think we can live without it.