Allow create_app to recognise config names
 ## Summary
`flask shell` is a neat feature provided by flask v1+ to allow dropping
into an application context from the command line. In order to do this,
we need to be able to construct an app from a string-representation of a
config class. This patch adds the ability to derive config class names
from strings.

In order to use `flask shell`, you just need to define `FLASK_APP` in
your `.env.` file in a fashion similar to this:

`FLASK_APP="application.factory:create_app('development')"`