Remove departmental references from the "Planned work" dashboard
Our measures Trello board tickets have names that begin with an internal
reference in square brackets, e.g. "[DWP 005] Unemployment".  These
references are useful internally for our analysts, but meaningless to
the wider public.  So we shouldn't show them on the public-facing
dashboard.

This uses a regular expression to match any leading text in square
brackets and remove it from the card title before sending it to be
displayed on the dashboard.