Show index colum header in table preview
The existing check for `== null` was not catching the fact that
index_column_caption is actually passed through as an empty string if
nothing has been entered in the form field.

By checking for "not index_column_caption" we will catch the case of
either null or an empty string, and so show the default header in the
preview.