Put category caption in second row of headers (if there is one)
In the on-page preview version of tables the category caption was shown
in the top header row, but in the published table it appears in the
second row of headers (if there is one).

This tweaks the logic for building headers to put the caption in the
second header row if there is one.

It also changes the top header row to use all <th> elements, rather than
<td> - this also matches what happens in the published table.