Create custom RDU WTForm fields
 ## Summary
This patch adds custom form fields for the WTForms library that will
allow us to more easily and consistently render forms in our templates
without having to worry so much about what fields we've put in the
forms. This will start to replace the `forms.html` macros template we
have. This work paves the way for the data source forms we'll use now
that we have them in a separate model from the main `page` table.

 ## Ticket
https://trello.com/c/G16YBvyq/1160