Use separate data source forms
 ## Summary
Data sources are now a separate model in our database, and the cleanest
way to integrate these into our big old measure page is to have a
separate form that can be used to ask for and collect data source
information.

This form will use the newly-added custom RDU WTForm fields to allow
easily rendering fields into our pages without having to duplicate the
field types in both the Python form and the HTML template. A main
`DataSourceForm` declares the fields, in line with the model, and
includes a compatability layer that allows it to write details back out
to the main `Page` table until a later PR removes it. We have a separate
`DataSource2Form` that subclasses the main form, because in our current
measure page content we include a distinct hint for the title field on
the secondary data source form. This is not particularly _nice_, but
it's the cleanest way I quickly thought of to implement the adjusted
content - if we really want to keep it.

All of the data source form fields have been removed from the
MeasurePageForm and its associated form, MeasurePageRequiredForm.

There is no associated DataSourceRequiredForm - I have instead opted to
validate this business requirement in another way. For context, when a
measure page is being worked on, there are a lot of fields, so it is
unreasonable to expect a user to provide data for all the fields in the
first session. Therefore, many fields are considered 'optional' during
the drafting process. However, when a user wants to submit the measure
for approval, we need to validate that all of the expected in a good
measure has been provided. The `MeasurePageRequiredForm` duplicates a
lot of the fields from the `MeasurePageForm`, and if we followed that
precedent we'd need to do something similar for the data source forms.
Instead, I've attached a custom validator to the DataSourceForm fields
that need to be present when the measure is submitted for review:
`RequiredForReviewValidator`. This will not raise a validation error
during the drafting process, but when the measure is submitted for
review will ensure data has been provided for the necessary fields.

After this commit is deployed, edits to measure pages will start to
duplicate changes to data sources in two places: the `Page` table, in
the existing data source fields, and the new `DataSource` table. This is
handled by the page service in the `_set_data_sources` and
`_update_measure_page_data_sources` methods. Once the migration is in,
and we're using the `DataSource` table exclusively, we can start to
remove these.

Copying measures or creating new versions of measure will duplicate the
data sources, even if no changes are made. This mimics the existing
functionality, and is something we should plan to address in the future,
once we consider whether or not there's a need or use for data sources
to be able to be common across measures.

One of the bugs fixed by this commit is the inability to totally remove
secondary sources. An additional field in the `DataSource` form,
`remove_data_source`, is toggled on by `_secondary_sources.js` when the
`Remove data source` link is clicked. If the form is submitted with this
value true, the data source will be deleted entirely, which is a clean
action now that it's modelled separately from the page. We also no
longer clear hidden fields on the page when removing sources: because
data sources have separate forms, each one contains a CSRF token in a
hidden field. Clearing the value of this field prevents the form from
validating when it's processed.

A new macro, `render_data_source_form`, encapsulates the logic for
rendering a `DataSourceForm`. This has been mostly simplified by the
custom RDU fields, and we should not be far away from only needing to do
`{{ data_source_form }}` in a template to render all the fields for it.
However, our "department select" box has not yet been modelled in a
custom RDU field. Once this has been done, we will be able to
dramatically simplify the rendering of forms in templates, and have
another easily-reused component in our arsenal.