Recognise "Updated version v3" label and use full organisation name
Analysts have added a new label "Updated version v3" that we want to
map to the same "Type" as "Updated version" for the planned pages
dashboard.

We also want to show the full organisation name rather than just the
acronym.  Because why would people understand the acronyms if they aren't
defined anywhere?

This also removes the RDU label from the list of organisations as we
don't have any statistics of our own.