Remove edit-and-preview measure page
 ## Summary
We aren't actively working on this, and there are no plans in the
immediate future, so maintaining it is a fair amount of wasted effort.
If we decide to start working on it again, we still have it in git
history, and indeed can just revert this commit. So let's burn it all
for now.

:fire: :fire: :fire: