Set timestamps on dimensions (#844)
This adds timestamps to the Dimension table to track the date and time that the dimension was first created, and when it was last updated.

These timestamps are necessary in order to give some feedback in the publisher interface on which dimensions have been "updated" since the last versions.

In contrast to the existing `created_at` and `updated_at` timestamps on the Page table, these timestamps are initially both set to the same 'current time' value by the Postgres database, not the Python application.  Subsequent updates also use the database’s clock to set the `updated_at` timestamp.

A data migration is also included to set the timestamps on the pre-existing dimensions. This sets the `created_at` timestamp on dimensions to be the same as the page it belongs to, and the `updated_at` timestamps on dimensions to either be the same as the `updated_at` timestamp on the page, or the `created_at` timestamp if it hasn't yet been updated.