Pull the full database using script
Currently the `pull_prod_data` command doesn't actually work, as it deliberately excludes the `users` table, which is required because the `user_page` table has foreign keys which depend on it.

The original reason for excluding the users table was so that you could maintain your local users, enabling you to still be able to login to your local version of the application. On balance though, I’m not sure this is workable.

This update pulls the full database, including the users table, the build table and the alembic_version table. All tables are also truncated before importing the data (in the right order).

After importing the data, the users table is "pseudonymised" to remove the e-mail addresses (replacing them with a random invalid one), the hashed passwords, and to make all users deactive.

To then gain access to the application you would need to create a new local user.