Remove unnecessary __init__ methods from "xxxRequiredForm"s
The `MeasurePageRequiredForm` and `DimensionRequiredForm` inherit from
`MeasurePageForm` and `DimensionForm` respectively.

The `__init__()` methods on these "Required" forms were identical to
that of the base classes, except for setting
`kwargs["meta"] = kwargs.get("meta") or {}`

This seems unnecessary as the only places where the "Required" forms are
created they are called with `meta` set anyway, so we can just fall
back to the `__init__()` of the base classes.