Allow data source removal by clearing fields
 ## Summary
This will allow users to remove data sources from a measure by
submitting the form with all fields blank. Note: the user cannot use
this method if they have selected a department for the source, as there
is no UI for blanking out this field. In this case, they'll need to use
the 'Remove data source' link.