BUGFIX: Duplicate rows at top of download file
We have seen an issue where the first 67 rows of a download file were
duplicated in the download.

It turns out that when downloading the file and reading row-by-row there
was no "UnicodeDecodeError" until line 68, where there was a `£` sign
for the first time.

Because the first 67 had successfully been read in to the `rows` array
as UTF-8, they were duplicated when re-reading the file as ISO-8859-1.

The simple fix here is to make sure the rows array is reset before
starting over.