Rename `_FormFieldSet`
 ## Summary
Rename the `_FormFieldSet` class to `_FormGroup`, which more accurately
describes what it does - it wraps one or more choice inputs in a form
group. Where there is more than one choice, it also includes a field
set, but that's more of an implementation detail.

(So is the name FormGroup, admittedly, but it also feels reasonably
generic enough to still work).