Speed up S3 copying via multithreading
 ## Summary
When using `pull_prod_data`, we also copy all S3 files from prod to our
own dev S3 buckets. There's no reason to do this synchronously and
linearly as our CPU is doing nothing in this process, so if we
multithread it we can wrap this stage up a whole lot faster.