Leave alembic_version alone when getting data from other environments
We shouldn't touch the `alembic_version` table when copying data between
environments. Doing so could result in an end state where the
`alembic_version` doesn't match the actual structure of the database,
which is potentialy a really annoying issue to track down if things
break because of it.

We should let Alembic handle structural database migrations and knowing
where it's got up to.