Replace "get_page_by_uri_and_type()" with "get_measure-page_hierarchy()"
When fetching measure pages the URL has the topic, subtopic and
measure URIs.

But if the view only checks the measure URI then we have what should be
invalid URLs (i.e. where the topic is not the parent of the subtopic, or
subtopic not the parent of the measure page) returning measure pages.

This change means that all parameters in the URL are matched when
fetching a measure page.