Apply context filter to handler, not logger
 ## Summary
We currently always log via our 'request handler', even when the log is
being emitted by something that isn't a request (e.g. management
commands). This adds an extra filter to the 'request logger handler' to
make sure that it's actually a request (i.e. has an IP address).
Otherwise, we don't log it.

This means that for management commands, we'll log via the default app
logger handler only (stdout with a different format).

We still patch in the contextual information about a request for all log
records.