Only use get_page_by_uri_and_type for topic and subtopic pages
It is possible to create measure pages with the same URI (in different
subtopics), and if that happens then the call to `.one()` here will
throw an error and we won't be able to get any page.

By restricting the method to topic and subtopic pages only we reduce the
risk of hitting this error.

Ideally "URI" would be enforced to be unique, in which case this method
would be fine.  This is work planned for soon, but in the meantime this
should help.