Lock build process to one at a time
 ## Summary
At the moment, we can have multiple processes building the static site
at the same time (in theory). Given that these write to the same S3
bucket, there are some race conditions that could result in an undefined
state at the end. This patch allows only one process to be building the
site at a time, which reduces this risk.

We might want to remove this if/when we start uploading builds to
separate S3 buckets and switching over for more atomic releases. But for
now, it should do the job.

 ## Ticket
https://trello.com/c/ZiEltV8m/1206