Incorporate leftover migration changes
 ## Summary
Whenever I generate a migration at the moment I'm getting a lot of
changes beyond the scope of what I have actually intended to apply. It
looks like we've somehow missed a few things in previous migrations.
This patch is a quick assessment of the migration that alembic wants to
apply, and a few tweaks to the models as I think are in line with what
we want, to make it so that new migrations will only contain the changes
actually added by a developer.