Remove duplicate foreign key constraint
 ## Summary
An historical migration is (slightly) broken because it creates a
foreign key constraint twice on the same table/column. Now that we have
a naming convention, it tries to create the second constraint with the
same name as the first, which causes everything to break.

This seems like an exceptional circumstance in which it's probably OK to
edit an historical migration.