Propose lightweight ADRs
 ## Summary
We need a way to record decisions (and the conversations around those
decisions) in a durable, discoverable format. Architecture Decision
Records (ADRs) are a way to record "significant" decisions made about
the technical architecture of our service, as well as the context around
those decisions, any alternatives that were considered or explored, and
the pros/cons that might be worth remembering.

We don't want to introduce a lot of complexity, so it seems beneficial
to keep this process very lightweight. In that vein, we will look at the
Lightweight Architecture Decision, as discussed here:

https://www.thoughtworks.com/radar/techniques/lightweight-architecture-decision-records
http://thinkrelevance.com/blog/2011/11/15/documenting-architecture-decisions

This does not involve an undue burden of detailed documentation, but
recommends preserving, concisely, the context of the problem, the
decision taken, the alternatives considered, and the consequences of the
chosen solution.

 ## Ticket
https://trello.com/c/ZiEltV8m/1206