First decision record for heroku scheduler
 ## Summary
The first lightweight 'architectural decision record' (ADR) that
preserves the context around a change being made to how we manage
building the static site.

At the moment, we have a set of scheduled tasks on Heroku that run
every 10 minutes, hour, or day. These range from building the
static site (every 10 minutes), to reporting stalled builds (every
hour), to refreshing the database from production data (in
dev/staging, every night). The main job here, building the static
site, has started to fail because it occasionally runs longer than
10 minutes. The Heroku scheduler will only allow a scheduled
instance to live for as long as the interval on the task. Heroku
is therefore killing the instance before it can finish, leading to
the incomplete builds and frequent reports of 'stalled' (in this
instance, killed) builds.

We need to schedule static site builds in a way that removes this
strict time constraint so that they will always finish building.

See ADR 0002 for further details.