Catch a NoResultFound from query on Page
Previously this method used `get_page_by_uri_and_type` for all page
types, and that raises a PageNotFoundException. Now we are querying the
model directly here for measure pages there's the possibility of a
NoResultFound from the cal to `.one()` - we have seen this recently in
the case where the uri for a new measure changes (because the title of
the page is updated), but the user has tried to access the old URL
(possibly via the back button on the browser).

This *does* raise the question of whether having infinitely changeable
uris for new measure pages is 100% a good thing.  On balance it's
probably better than being stuck with the first thing that gets typed
into the Title field though.