Remove UK from UKCountry enum
 ## Summary
The 'UK' is not a country in the UK. This patch removes UK from the
enum, and includes a migration that expands 'UK' to
'ENGLAND+WALES+SCOTLAND+NORTHERN_IRELAND' in the database. This will
allow a future migration to replace the existing enum (which will still
contain 'UK', but unused) with a version that doesn't contain UK. We
need to do this as a separate migration to ensure code is backwards
compatible with existing app instances.

It also removes just-added functionality from the `RDUCheckboxField`,
specifically the `exclude_enum_fields` parameter, which was added solely
to support the UKCountry enum being used in the MeasurePageForm. We
didn't want to show the UK field to users, so needed a way to prevent it
populating into the form. Now that UK is no longer in the enum, we don't
need this, so for the sake of simplicity, it's been removed.

Also removed is the CHECKBOX_ENUM_LOOKUPS dict, which seems to no longer
have any references pointing at it.