Don't import the same thing from different places
We were importing a bunch of things from `sqlalchemy.orm` as well as
importing `db` from `application` which *is* an SQLAlchemy object.

And then sometimes we had `relationship` (or the outdated `relation`)
and in other places called the same thing as `db.relationship`.

It seems neater to use our `db` object as the thing to import and call
methods on that, rather than import from `sqlalchemy.orm` directly, so
that's what I've done here.