Make consistent with static site build
 ## Summary
This `if` check should not be necessary, as the helper method performs a
conditional inside it. Also, the static site build
`build.py:write_measure_page` doesn't do an external conditional, so
this at least makes us more consistent and presents the same information
internally as we currently show externally.