Remove old views from main code directory
 ## Summary
We needed to have separate versions of the materialized views that
migrations could reference as immutable records of how the views looked
at that point in time. At the time we decided to create a copy of the
file next to the original and then import from there, but this creates
duplication and confusion in the codebase - one file contains live code,
and the other a static reference to a point in time.

It feels cleaner to inline the static references in the migration files
themselves, and then only have the live materialized view queries in the
main codebase.